Rename BlackBox and WhiteBox into OperatorBox and BlueprintBox
