+++ src/slang/ui/views/blueprint.ts
@@ -5,9 +5,9 @@ import {GenericPortModel, PortModel} from "../../core/abstract/port";
 import {SlangSubject} from "../../core/abstract/utils/events";
 import {BlueprintModel} from "../../core/models";
 import {TypeIdentifier} from "../../definitions/type";
-import {BlackBoxShape} from "../components/blackbox";
+import {BlackBoxShape} from "../components/operator";
 import {ConnectionElement} from "../components/connection";
-import {BlueprintBox} from "../components/whitebox";
+import {BlueprintBox} from "../components/blueprint";
 import {ViewFrame} from "../frame";
 import {PaperView, PaperViewArgs} from "./paper-view";
 import {TargetableComponent, UserEvent } from "./user-events";
